requirements.txt
requests>=2.31.0
aiohttp>=3.9.1
pytz>=2024.1
slack-sdk>=3.26.1
openai>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
cairosvg>=2.7.0
aiofiles>=23.2.1